requirements.txt
einops==0.4.1
matplotlib==3.3.4
numpy==1.20.1
opencv_python==4.5.5.64
Pillow==9.0.1
PyYAML==6.0
scipy==1.6.2
streamlit==1.7.0
tensorflow==2.5.0
tensorflow_addons==0.16.1
timm==0.5.4
torch==1.7.1+cu110
yacs==0.1.8
